requirements.txt
backports.shutil-get-terminal-size==1.0.0
bottle==0.12.9
decorator==4.0.10
ipython==4.2.0
ipython-genutils==0.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.4
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
telepot==8.3
traitlets==4.2.2
urllib3==1.9.1
wcwidth==0.1.7
APScheduler==3.2.0
feedparser==5.2.1
redis==2.10.5
pytz==2016.6.1
python-dateutil==2.5.3